LPS-54184 we seem to use *Adapter more than *Adaptor
