+++ modules/portal/portal-wab-extender/src/com/liferay/portal/wab/extender/internal/adapter/HttpAdapter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.wab.extender.internal.adaptor;
+package com.liferay.portal.wab.extender.internal.adapter;
 
 import com.liferay.portal.kernel.util.HashMapDictionary;
